% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMotifFromTxt.R
\name{getMotifFromTxt}
\alias{getMotifFromTxt}
\title{Extract and Visualize Motif Information from MEME Software}
\arguments{
\item{data}{A txt file from MEME software.}
}
\value{
Return a vector or a datafram
}
\description{
\code{getMotifFromXML} Extract motif information from the MEME software results, .txt file.
}
\examples{
filepath <- system.file("examples", "meme.txt", package = "ggmotif")
motif_extract <- getMotifFromTxt(data = filepath)
}
\author{
Xiang LI <lixiang117423@foxmail.com>
}
