% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMotifFromXML.R
\name{getMotifFromXML}
\alias{getMotifFromXML}
\title{Extract and Visualize Motif Information from MEME Software}
\arguments{
\item{XMLfile}{A XML file from MEME software.}
}
\value{
Return a vector or a datafram
}
\description{
\code{getMotifFromXML} Extract motif information from the MEME software results, XML file.
}
\examples{
filepath <- system.file("examples", "meme.xml", package="ggmotif")
motif_extract <- getMotifFromXML(XMLfile = filepath)
}
\author{
Xiang LI <lixiang117423@foxmail.com>
}
