% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggmotif.R, R/motif_plot.R
\docType{package}
\name{ggmotif}
\alias{ggmotif}
\alias{ggmotif-package}
\title{ggmotif}
\usage{
ggmotif(x, x.pos = 0, y.pos = 0, ht = 1, wt = 1, id = "")
}
\arguments{
\item{x}{either a character string, a \code{\link[Biostrings]{DNAString}} object, or a matrix}

\item{x.pos}{position on x axis for bottom left corner of motif}

\item{y.pos}{position on y axis for bottom left corner of motif}

\item{ht}{height of motif, default is 1}

\item{wt}{width of each letter of motif, default is 1}

\item{id}{id}
}
\value{
geom_polygon to be added to ggplot. this will also add an ugly
legend... to remove, use \code{\link{ggmotif_scale}}
}
\description{
package to help plot DNA sequence motifs on ggplots

ggmotif_plot
}
\examples{
ggplot() + ggmotif("ACG",x.pos = 5,y.pos = 2) + ggmotif_scale()
}

