% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_plot.R
\name{ggmotif_plot}
\alias{ggmotif_plot}
\title{ggmotif_plot}
\usage{
ggmotif_plot(x, y.axis = FALSE, y.lab = "Letter Frequency", x.axis = TRUE,
  x.lab = "")
}
\arguments{
\item{x}{either a character string, a \code{\link[Biostrings]{DNAString}} object, or a matrix}

\item{y.axis}{plot a y axis? default is FALSE}

\item{y.lab}{axis title for y axis, if y.axis is TRUE}

\item{x.axis}{plot a x axis? default is TRUE}

\item{x.lab}{title for x axis, if x.axis is TRUE}
}
\value{
ggplot with motif
}
\description{
ggmotif_plot
}
\examples{
ggmotif_plot("ACGTRWN")
ggmotif_plot(Biostrings::DNAString("ACGTRWN"))
ggmotif_plot(Biostrings::consensusMatrix("ACGTRWN"))
}

