% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/motif_plot.R
\name{ggmotif_scale}
\alias{ggmotif_scale}
\title{ggmotif_scale}
\usage{
ggmotif_scale()
}
\description{
removes ugly legends and applies fixed coloring scheme to letters
}
\examples{
ggplot() + ggmotif("ACG") + ggmotif_scale()
}

