% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{ggmotif_theme}
\alias{ggmotif_theme}
\title{ggmotif_theme}
\usage{
ggmotif_theme(base_size = 8, base_family = "Helvetica")
}
\arguments{
\item{base_size}{font size}

\item{base_family}{font family}
}
\description{
theme for use with ggmotif
}
\examples{

ggplot() + ggmotif("ACG",x.pos = 5,y.pos = 2) + ggmotif_scale() + ggmotif_theme()
}

