% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_location.R
\name{motif_location}
\alias{motif_location}
\title{Extract and Visualize Motif Information from MEME Software}
\arguments{
\item{data}{A data frame file from getMotifFromXML function.}
}
\value{
Return a datafram
}
\description{
\code{getMotifFromXML} Visualize motif location in a specificial sequences..
}
\examples{
filepath <- system.file("examples", "meme.xml", package="ggmotif")
motif_extract <- getMotifFromXML(XMLfile = filepath)
motif_plot <- motif_location(data = motif_extract)

}
\author{
Xiang LI <lixiang117423@foxmail.com>
}
