% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_with_tree.R
\name{motif_with_tree}
\alias{motif_with_tree}
\title{Extract and Visualize Motif Information from MEME Software}
\arguments{
\item{data}{A data frame file from getMotifFromXML function.}

\item{tree}{A tree file.}
}
\value{
Return a datafram
}
\description{
\code{getMotifFromXML} Visualize motif location with a phylogenetic tree object..
}
\examples{
filepath <- system.file("examples", "meme.xml", package="ggmotif")
treepath <- system.file("examples", "tree.nwk", package="ggmotif")
motif_extract <- getMotifFromXML(XMLfile = filepath)
motif_plot_with_tree <- motif_with_tree(data = motif_extract, tree = treepath)

}
\author{
Xiang LI <lixiang117423@foxmail.com>
}
